requirements.txt
Django==1.9.1
Pillow==3.1.0
argparse==1.2.1
django-autofixture==0.11.0
django-tinymce==2.2.0
sorl-thumbnail==12.3
wsgiref==0.1.2
